I've refactored `GameInterface.tsx` for improved modularity and maintainability.
Here are the key changes:
- I extracted the core game logic and state management into a new custom hook called `useNBackLogic.ts`.
- I split the UI of `GameInterface.tsx` into three distinct components based on the game state:
    - `GameSetupScreen.tsx`
    - `GamePlayScreen.tsx`
    - `GameResultsScreen.tsx`
- `GameInterface.tsx` now acts as a container, using `useNBackLogic` and showing the right component depending on the situation.
- I moved game-related type definitions (`GameMode`, `GameState`, `GameSession`) to `src/types/game.ts`.
- I updated the existing tests in `GameInterface.test.tsx` to match the new structure, including setting up mocks for `useNBackLogic` and `localStorage`. All tests are currently passing.

This refactoring makes `GameInterface.tsx` much simpler, which in turn makes your codebase easier to understand, test, and maintain.
